+++ CHANGELOG.md
@@ -1,1 +1,17 @@
-# Changelog
+# Changelog for unreleased
+
+The following sections list the changes for unreleased.
+
+## Summary
+
+ * Chg #3: Initial release of basic version
+
+## Details
+
+ * Change #3: Initial release of basic version
+
+   Just prepared an initial basic version which could be released to the public.
+
+   https://github.com/webhippie/terrastate/issues/3
+
+